poogie: Add core poogie processing module
This module takes care of type decoding and generates a list of events
which can then be processed.

It comes with a companion script get-quests.py which uses poogie to
generate some data about quests, such as quest name, hunters (and
their equipment and faint count), and so forth. Most of this
functionality should move into poogie itself at some point.